requirements.txt
# PyTorch (CPU). For CUDA replace 'cpu' with 'cu121' or 'cu124'
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.11.0+cpu
torchvision==0.26.0+cpu

# Core
numpy==2.4.4
opencv-python==4.13.0.92
Pillow==12.2.0
matplotlib==3.10.8
click==8.3.2
marshmallow==4.3.0
marshmallow-dataclass==8.7.1
joblib==1.5.3
PyYAML==6.0.3
reportlab==4.4.10
